+++ plugins/csp-stars/src/Plugin.cpp
@@ -233,20 +233,13 @@ void Plugin::deInit() {
 void Plugin::update() {
 
   // Update the stars brightness based on the scene's pApproximateSceneBrightness. This is to fade
-  // out the stars when we are close to a Planet. If HDR rendering is enabled, we will not change
-  // the star's brightness.
-  float fIntensity = 1.F - mGraphicsEngine->pApproximateSceneBrightness.get();
-
-  if (mAllSettings->mGraphics.pEnableHDR.get()) {
-    fIntensity = 1.F;
-  }
-
-  mStars->setLuminanceMultiplicator(
-      fIntensity * std::exp(mPluginSettings.mLuminanceMultiplicator.get()));
-  mStars->setCelestialGridColor(VistaColor(0.5F, 0.8F, 1.F,
-      0.3F * fIntensity * (mPluginSettings.mEnableCelestialGrid.get() ? 1.F : 0.F)));
-  mStars->setStarFiguresColor(VistaColor(
-      0.5F, 1.F, 0.8F, 0.3F * fIntensity * (mPluginSettings.mEnableStarFigures.get() ? 1.F : 0.F)));
+  // out the stars when we are close to a planet in non-HDR mode.
+  mStars->setApproximateSceneBrigthness(mGraphicsEngine->pApproximateSceneBrightness.get());
+  mStars->setLuminanceMultiplicator(std::exp(mPluginSettings.mLuminanceMultiplicator.get()));
+  mStars->setCelestialGridColor(
+      VistaColor(0.5F, 0.8F, 1.F, 0.3F * (mPluginSettings.mEnableCelestialGrid.get() ? 1.F : 0.F)));
+  mStars->setStarFiguresColor(
+      VistaColor(0.5F, 1.F, 0.8F, 0.3F * (mPluginSettings.mEnableStarFigures.get() ? 1.F : 0.F)));
 
   auto mat = mSolarSystem->getObject("Barycenter")->getObserverRelativeTransform();