+++ plugins/csp-trajectories/src/DeepSpaceDot.cpp
@@ -8,7 +8,7 @@
 #include "DeepSpaceDot.hpp"
 
 #include "../../../src/cs-core/SolarSystem.hpp"
-#include "../../../src/cs-utils/FrameTimings.hpp"
+#include "../../../src/cs-utils/FrameStats.hpp"
 #include "../../../src/cs-utils/utils.hpp"
 
 #include <VistaKernel/GraphicsManager/VistaGraphicsManager.h>
@@ -145,7 +145,7 @@ bool DeepSpaceDot::Do() {
     return true;
   }
 
-  cs::utils::FrameTimings::ScopedTimer timer("Dot of " + mObjectName);
+  cs::utils::FrameStats::ScopedTimer timer("Dot of " + mObjectName);
   // get viewport to draw dot with correct aspect ration
   std::array<GLint, 4> viewport{};
   glGetIntegerv(GL_VIEWPORT, viewport.data());