+++ src/cs-core/GraphicsEngine.cpp
@@ -29,6 +29,44 @@ namespace cs::core {
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 
+void GLAPIENTRY MessageCallback(GLenum source, GLenum type, GLuint id, GLenum severity,
+    GLsizei length, const GLchar* message, const void* userParams) {
+
+  // get the log level from the settings
+  const cs::core::Settings* settings = reinterpret_cast<const cs::core::Settings*>(userParams);
+
+  // Print the following infos (OpenGL errors, shader compile errors, perf. warnings, shader
+  // compilation warnings, depricated code, redundant state changes, undefined behaviour, anything
+  // that isnt an error or perf. issue)
+  if (settings->pLogLevelGL.get() <= spdlog::level::debug &&
+      severity == GL_DEBUG_SEVERITY_NOTIFICATION) {
+    logger().debug("{}", message);
+    return;
+  }
+
+  // Print the following infos (OpenGL errors, shader compile errors, perf. warnings, shader
+  // compilation warnings, depricated code, redundant state changes, undefined behaviour)
+  if (settings->pLogLevelGL.get() <= spdlog::level::info && severity == GL_DEBUG_SEVERITY_LOW) {
+    logger().info("{}", message);
+    return;
+  }
+
+  // Print the following infos (OpenGL errors, shader compile errors, perf. warnings, shader
+  // compilation warnings, depricated code)
+  if (settings->pLogLevelGL.get() <= spdlog::level::warn && severity == GL_DEBUG_SEVERITY_MEDIUM) {
+    logger().warn("{}", message);
+    return;
+  }
+
+  // Print the following infos (OpenGL errors, shader compile errors)
+  if (settings->pLogLevelGL.get() <= spdlog::level::critical &&
+      severity == GL_DEBUG_SEVERITY_HIGH) {
+    logger().error("{}", message);
+  }
+}
+
+////////////////////////////////////////////////////////////////////////////////////////////////////
+
 GraphicsEngine::GraphicsEngine(std::shared_ptr<core::Settings> settings)
     : mSettings(std::move(settings))
     , mShadowMap(std::make_shared<graphics::ShadowMap>())
@@ -50,6 +88,19 @@ GraphicsEngine::GraphicsEngine(std::shared_ptr<core::Settings> settings)
         ->second->GetWindowProperties()
         ->SetVSyncEnabled(value);
   });
+
+  // setup OpenGL debugging ------------------------------------------------------------------------
+
+  mSettings->pLogLevelGL.connectAndTouch([](auto level) {
+    if (level == spdlog::level::off) {
+      glDisable(GL_DEBUG_OUTPUT);
+    } else {
+      glEnable(GL_DEBUG_OUTPUT);
+    }
+  });
+
+  // Attach the debug callback to print the messages.
+  glDebugMessageCallback(MessageCallback, static_cast<void*>(mSettings.get()));
 
   // setup shadows ---------------------------------------------------------------------------------
 
@@ -255,39 +306,6 @@ std::shared_ptr<graphics::HDRBuffer> GraphicsEngine::getHDRBuffer() const {
 std::vector<std::shared_ptr<graphics::EclipseShadowMap>> const&
 GraphicsEngine::getEclipseShadowMaps() const {
   return mEclipseShadowMaps;
-}
-
-////////////////////////////////////////////////////////////////////////////////////////////////////
-
-bool glDebugOnlyErrors = true;
-
-void GLAPIENTRY oglMessageCallback(GLenum /*source*/, GLenum type, GLuint /*id*/, GLenum severity,
-    GLsizei /*length*/, const GLchar* message, const void* /*userParam*/) {
-
-  if (type == GL_DEBUG_TYPE_ERROR || GL_DEBUG_TYPE_UNDEFINED_BEHAVIOR) {
-    if (severity == GL_DEBUG_SEVERITY_HIGH) {
-      logger().critical(message);
-    } else {
-      logger().error(message);
-    }
-  } else if (!glDebugOnlyErrors) {
-    if (severity == GL_DEBUG_SEVERITY_NOTIFICATION) {
-      logger().debug(message);
-    } else {
-      logger().warn(message);
-    }
-  }
-}
-
-void GraphicsEngine::enableGLDebug(bool onlyErrors) {
-  glDebugOnlyErrors = onlyErrors;
-  glEnable(GL_DEBUG_OUTPUT);
-  glDebugMessageCallback(oglMessageCallback, nullptr);
-}
-
-void GraphicsEngine::disableGLDebug() {
-  glDisable(GL_DEBUG_OUTPUT);
-  glDebugMessageCallback(nullptr, nullptr);
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////