+++ src/cs-graphics/ToneMappingNode.cpp
@@ -7,7 +7,7 @@
 
 #include "ToneMappingNode.hpp"
 
-#include "../cs-utils/FrameTimings.hpp"
+#include "../cs-utils/FrameStats.hpp"
 #include "HDRBuffer.hpp"
 
 #include <VistaInterProcComm/Cluster/VistaClusterDataCollect.h>
@@ -465,7 +465,7 @@ float ToneMappingNode::getLastMaximumLuminance() const {
 
 bool ToneMappingNode::ToneMappingNode::Do() {
 
-  utils::FrameTimings::ScopedTimer timer("Tonemapping");
+  utils::FrameStats::ScopedTimer timer("Tonemapping");
 
   if (mShaderDirty) {